+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ## v0.10.0 - dev
 
+- **(breaking)** `StabMixture` was renamed to `GeneralizedStabilizer`.
 - Constructing a `Destabilizer` out of a full-rank `Stabilizer` does not require a canonicalization anymore, i.e. `stabilizerview(Destabilizer(s))==s` is guaranteed.
 - The `maximally_mixed` function is now available for creating maximally mixed multi-qubit states.
 - Gate errors are now conveniently supported by the various ECC benchmark setups in the `ECC` module.